+++ src/main/java/lab2ass/OnlineCourse.java
@@ -1,12 +1,12 @@
 package lab2ass;
 
 public class OnlineCourse extends Course {
-	public OnlineCourse (int _studentID, String _name, float _price, int _runtime) {
+	public OnlineCourse (int _studentID, String _name, double _price, int _runtime) {
 		super(_studentID, _name, _price, _runtime);
 	}
 
 	@Override
-	public float getCost() {
+	public double getCost() {
 		return 0;
 	}
 }
