+++ src/main/java/lab2ass/ShortCourse.java
@@ -1,7 +1,7 @@
 package lab2ass;
 
 public class ShortCourse extends Course {
-	public ShortCourse (int _studentID, String _name, float _price, int _runtime) {
+	public ShortCourse (int _studentID, String _name, double _price, int _runtime) {
 		super(_studentID, _name, _price, _runtime);
 	}
 }
